Extract `WrapTriangle3d` and `WrapTrianglePad` to `wrap_polygon.cpp`
* Move the C++ wrapper classes `WrapTriangle3d` and `WrapTrianglePad` from file
  `wrap_World.cpp` to `wrap_polygon.cpp`.

Claude code command:
* Extract `WrapTriangle3d` and `WrapTrianglePad` from `wrap_World.cpp` to
  `wrap_polygon.cpp`
